refactor: updated validation for cli
